deploy/requirements.txt
fastapi
uvicorn[standard]
torch
Pillow
ultralytics
